+++ CnCTDRAMapEditor/RedAlert/GamePluginRA.cs
@@ -651,7 +651,7 @@ namespace MobiusEditor.RedAlert
                 TerrainTypes.GetTypes(), OverlayTypes.GetTypes(), SmudgeTypes.GetTypes(Globals.ConvertCraters),
                 EventTypes.GetTypes(), cellEventTypes, unitEventTypes, structureEventTypes, terrainEventTypes,
                 ActionTypes.GetTypes(), cellActionTypes, unitActionTypes, structureActionTypes, terrainActionTypes,
-                MissionTypes.GetTypes(), MissionTypes.MISSION_GUARD, MissionTypes.MISSION_STOP, MissionTypes.MISSION_HARVEST,
+                MissionTypes.GetTypes(), MissionTypes.GetUnassignableTypes(), MissionTypes.MISSION_GUARD, MissionTypes.MISSION_STOP, MissionTypes.MISSION_HARVEST,
                 MissionTypes.MISSION_UNLOAD, DirectionTypes.GetMainTypes(), DirectionTypes.GetAllTypes(), InfantryTypes.GetTypes(),
                 UnitTypes.GetTypes(Globals.DisableAirUnits), BuildingTypes.GetTypes(), TeamMissionTypes.GetTypes(),
                 fullTechnoTypes, waypoints, movieTypes, movieEmpty, themeEmpty.Yield().Concat(themeTypes), themeEmpty,
@@ -1616,9 +1616,9 @@ namespace MobiusEditor.RedAlert
                 int dirValue;
                 if (!Int32.TryParse(tokens[4], out dirValue))
                 {
-                    errors.Add(string.Format("Direction for unit '{0}' on cell {1}, value '{2}', cannot be parsed; skipping.", unitType.Name, cell, tokens[4]));
+                    errors.Add(string.Format("Direction for unit '{0}' on cell {1}, value '{2}', cannot be parsed; clearing to 0.", unitType.Name, cell, tokens[4]));
                     modified = true;
-                    continue;
+                    dirValue = 0;
                 }
                 DirectionType dirType = DirectionType.FindClosestDirectionType(dirValue, Map.UnitDirectionTypes);
                 if (dirType.ID != dirValue)
@@ -1633,7 +1633,7 @@ namespace MobiusEditor.RedAlert
                     House = Map.HouseTypes.Where(t => t.Equals(tokens[0])).FirstOrDefault(),
                     Strength = strength,
                     Direction = dirType,
-                    Mission = Map.MissionTypes.Where(t => t.Equals(tokens[5])).FirstOrDefault() ?? Map.GetDefaultMission(unitType),
+                    Mission = Map.MissionTypes.Where(t => t.Equals(tokens[5])).FirstOrDefault(),
                 };
                 if (newUnit.House == null)
                 {
@@ -1650,6 +1650,15 @@ namespace MobiusEditor.RedAlert
                     }
                     modified = true;
                     newUnit.House = defHouse;
+                }
+                if (newUnit.Mission == null)
+                {
+                    string badOrder = Map.MissionTypesBad.Where(t => t.Equals(tokens[5], StringComparison.InvariantCultureIgnoreCase)).FirstOrDefault();
+                    string deforders = Map.GetDefaultMission(unitType);
+                    string reason = badOrder != null ? "unsupported" : "unknown";
+                    errors.Add(string.Format("Unit '{0}' on cell {1} references {2} orders '{3}'; changing to '{4}'.",
+                        newUnit.Type.Name, cell, reason, badOrder ?? tokens[5], deforders));
+                    newUnit.Mission = deforders;
                 }
                 if (!Map.Technos.Add(cell, newUnit))
                 {
@@ -1773,10 +1782,10 @@ namespace MobiusEditor.RedAlert
                 int dirValue;
                 if (!Int32.TryParse(tokens[4], out dirValue))
                 {
-                    errors.Add(string.Format("Direction for aircraft '{0}' on cell {1}, value '{2}', cannot be parsed; skipping.",
+                    errors.Add(string.Format("Direction for aircraft '{0}' on cell {1}, value '{2}', cannot be parsed; clearing to 0.",
                         aircraftType.Name, cell, tokens[4]));
                     modified = true;
-                    continue;
+                    dirValue = 0;
                 }
                 DirectionType dirType = DirectionType.FindClosestDirectionType(dirValue, Map.UnitDirectionTypes);
                 if (dirType.ID != dirValue)
@@ -1785,31 +1794,40 @@ namespace MobiusEditor.RedAlert
                         aircraftType.Name, cell, dirValue, dirType.ID, dirType.Name));
                     modified = true;
                 }
-                Unit newUnit = new Unit()
+                Unit newAir = new Unit()
                 {
                     Type = aircraftType,
                     House = Map.HouseTypes.Where(t => t.Equals(tokens[0])).FirstOrDefault(),
                     Strength = strength,
                     Direction = dirType,
-                    Mission = Map.MissionTypes.Where(t => t.Equals(tokens[5])).FirstOrDefault() ?? Map.GetDefaultMission(aircraftType)
+                    Mission = Map.MissionTypes.Where(t => t.Equals(tokens[5])).FirstOrDefault(),
                 };
-                if (newUnit.House == null)
+                if (newAir.House == null)
                 {
                     HouseType defHouse;
                     if ("ITALY".Equals(tokens[0], StringComparison.OrdinalIgnoreCase))
                     {
                         defHouse = HouseTypes.Ukraine;
-                        errors.Add(string.Format("Aircraft '{0}' on cell {1} has obsolete house '{2}'; substituting with '{3}'.", newUnit.Type.Name, cell, tokens[0], defHouse.Name));
+                        errors.Add(string.Format("Aircraft '{0}' on cell {1} has obsolete house '{2}'; substituting with '{3}'.", newAir.Type.Name, cell, tokens[0], defHouse.Name));
                     }
                     else
                     {
                         defHouse = Map.HouseTypes.First();
-                        errors.Add(string.Format("Aircraft '{0}' on cell {1} references unknown house '{2}'; clearing to '{3}'.", newUnit.Type.Name, cell, tokens[0], defHouse.Name));
+                        errors.Add(string.Format("Aircraft '{0}' on cell {1} references unknown house '{2}'; clearing to '{3}'.", newAir.Type.Name, cell, tokens[0], defHouse.Name));
                     }
                     modified = true;
-                    newUnit.House = defHouse;
+                    newAir.House = defHouse;
                 }
-                if (!Map.Technos.Add(cell, newUnit))
+                if (newAir.Mission == null)
+                {
+                    string badOrder = Map.MissionTypesBad.Where(t => t.Equals(tokens[5], StringComparison.InvariantCultureIgnoreCase)).FirstOrDefault();
+                    string deforders = Map.GetDefaultMission(aircraftType);
+                    string reason = badOrder != null ? "unsupported" : "unknown";
+                    errors.Add(string.Format("Aircraft '{0}' on cell {1} references {2} orders '{3}'; changing to '{4}'.",
+                        newAir.Type.Name, cell, reason, badOrder ?? tokens[5], deforders));
+                    newAir.Mission = deforders;
+                }
+                if (!Map.Technos.Add(cell, newAir))
                 {
                     ICellOccupier techno = Map.Technos[cell];
                     if (techno is Building building)
@@ -1901,9 +1919,9 @@ namespace MobiusEditor.RedAlert
                 int dirValue;
                 if (!Int32.TryParse(tokens[4], out dirValue))
                 {
-                    errors.Add(string.Format("Direction for ship '{0}' on cell {1}, value '{2}', cannot be parsed; skipping.", vesselType.Name, cell, tokens[4]));
+                    errors.Add(string.Format("Direction for ship '{0}' on cell {1}, value '{2}', cannot be parsed; clearing to 0.", vesselType.Name, cell, tokens[4]));
                     modified = true;
-                    continue;
+                    dirValue = 0;
                 }
                 DirectionType dirType = DirectionType.FindClosestDirectionType(dirValue, Map.UnitDirectionTypes);
                 if (dirType.ID != dirValue)
@@ -1918,7 +1936,7 @@ namespace MobiusEditor.RedAlert
                     House = Map.HouseTypes.Where(t => t.Equals(tokens[0])).FirstOrDefault(),
                     Strength = strength,
                     Direction = dirType,
-                    Mission = Map.MissionTypes.Where(t => t.Equals(tokens[5])).FirstOrDefault() ?? Map.GetDefaultMission(vesselType),
+                    Mission = Map.MissionTypes.Where(t => t.Equals(tokens[5])).FirstOrDefault(),
                 };
                 if (newShip.House == null)
                 {
@@ -1935,6 +1953,15 @@ namespace MobiusEditor.RedAlert
                     }
                     modified = true;
                     newShip.House = defHouse;
+                }
+                if (newShip.Mission == null)
+                {
+                    string badOrder = Map.MissionTypesBad.Where(t => t.Equals(tokens[5], StringComparison.InvariantCultureIgnoreCase)).FirstOrDefault();
+                    string deforders = Map.GetDefaultMission(vesselType);
+                    string reason = badOrder != null ? "unsupported" : "unknown";
+                    errors.Add(string.Format("Ship '{0}' on cell {1} references {2} orders '{3}'; changing to '{4}'.",
+                        newShip.Type.Name, cell, reason, badOrder ?? tokens[5], deforders));
+                    newShip.Mission = deforders;
                 }
                 if (!Map.Technos.Add(cell, newShip))
                 {
@@ -2092,10 +2119,10 @@ namespace MobiusEditor.RedAlert
                 int dirValue;
                 if (!Int32.TryParse(tokens[6], out dirValue))
                 {
-                    errors.Add(string.Format("Direction for infantry '{0}' on cell {1}, sub-position {2}, value '{3}', cannot be parsed; skipping.",
+                    errors.Add(string.Format("Direction for infantry '{0}' on cell {1}, sub-position {2}, value '{3}', cannot be parsed; clearing to 0.",
                         infantryType.Name, cell, stoppingPos, tokens[6]));
                     modified = true;
-                    continue;
+                    dirValue = 0;
                 }
                 DirectionType dirType = DirectionType.FindClosestDirectionType(dirValue, Map.UnitDirectionTypes);
                 if (dirType.ID != dirValue)
@@ -2133,7 +2160,7 @@ namespace MobiusEditor.RedAlert
                     House = Map.HouseTypes.Where(t => t.Equals(tokens[0])).FirstOrDefault(),
                     Strength = strength,
                     Direction = dirType,
-                    Mission = Map.MissionTypes.Where(t => t.Equals(tokens[5])).FirstOrDefault() ?? Map.GetDefaultMission(infantryType),
+                    Mission = Map.MissionTypes.Where(t => t.Equals(tokens[5])).FirstOrDefault(),
                     Trigger = tokens[7]
                 };
                 infantryGroup.Infantry[stoppingPos] = inf;
@@ -2152,6 +2179,15 @@ namespace MobiusEditor.RedAlert
                     }
                     modified = true;
                     inf.House = defHouse;
+                }
+                if (inf.Mission == null)
+                {
+                    string badOrder = Map.MissionTypesBad.Where(t => t.Equals(tokens[5], StringComparison.InvariantCultureIgnoreCase)).FirstOrDefault();
+                    string deforders = Map.GetDefaultMission(infantryType);
+                    string reason = badOrder != null ? "unsupported" : "unknown";
+                    errors.Add(string.Format("Infantry '{0}' on cell {1}, sub-position {2} references {3} orders '{4}'; changing to '{5}'.",
+                        inf.Type.Name, cell, stoppingPos, reason, badOrder ?? tokens[5], deforders));
+                    inf.Mission = deforders;
                 }
             }
         }
@@ -2228,10 +2264,10 @@ namespace MobiusEditor.RedAlert
                 int dirValue;
                 if (!Int32.TryParse(tokens[4], out dirValue))
                 {
-                    errors.Add(string.Format("Direction for structure '{0}' on cell {1}, value '{2}', cannot be parsed; skipping.",
+                    errors.Add(string.Format("Direction for structure '{0}' on cell {1}, value '{2}', cannot be parsed; clearing to 0.",
                         buildingType.Name, cell, tokens[4]));
                     modified = true;
-                    continue;
+                    dirValue = 0;
                 }
                 DirectionType dirType = DirectionType.FindClosestDirectionType(dirValue, Map.BuildingDirectionTypes);
                 if (dirType.ID != dirValue)
@@ -2469,17 +2505,9 @@ namespace MobiusEditor.RedAlert
                 };
                 if (!Map.Technos.Add(cell, newTerr))
                 {
-                    ICellOccupier techno = Map.FindBlockingObject(cell, terrainType, out int blockingCell, out int placementCell);
+                    ICellOccupier techno = Map.FindBlockingObject(cell, terrainType, false, out int blockingCell, out int placementCell);
                     string reportCell = blockingCell == -1 ? "<unknown>" : blockingCell.ToString();
-                    if (techno is Building building)
-                    {
-                        errors.Add(string.Format("Terrain '{0}' on cell {1} overlaps structure '{2}' placed on cell {3} at cell {4}; skipping.", terrainType.Name, cell, building.Type.Name, placementCell, reportCell));
-                    }
-                    else if (techno is Overlay overlay)
-                    {
-                        errors.Add(string.Format("Terrain '{0}' on cell {1} overlaps overlay '{2}' at cell {3}; skipping.", terrainType.Name, cell, overlay.Type.Name, reportCell));
-                    }
-                    else if (techno is Terrain terrain)
+                    if (techno is Terrain terrain)
                     {
                         errors.Add(string.Format("Terrain '{0}' on cell {1} overlaps terrain '{2}' placed on cell {3} at cell {4}; skipping.", terrainType.Name, cell, terrain.Type.Name, placementCell, reportCell));
                     }